feat: allow amd64 binaries of pd, gonut & yft to install on M1
M1 Macs come with Rosetta emulation which is able to run binaries
built for amd64 to run. This ofcourse impacts the performance but
until native binaries are built this would help unblock installation
of other binaries on M1.

Without this brew tap homeport/tap fails beause there is not url
for pd/yft/gonut.

Fixes: https://github.com/homeport/homebrew-tap/issues/52